requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
tensorflow>=2.6.0
pillow>=8.3.1
python-multipart
numpy>=1.19.5